+++ src/newamp1.c
@@ -497,7 +497,7 @@ void newamp1_model_to_indexes(C2CONST *c2const,
 
     /* running sum of squared error for variance calculation */
     for(k=0; k<K; k++)
-        *se += pow(rate_K_vec_no_mean[k]-rate_K_vec_no_mean_[k],2.0);
+        *se += (float)pow(rate_K_vec_no_mean[k]-rate_K_vec_no_mean_[k],2.0);
 
     /* scalar quantise mean (effectively the frame energy) */
     float w[1] = {1.0};
